backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pandas
openai>=1.35.0
python-dotenv
